Response format changed

+++ src/main/java/com/example/talent_api/controllers/HiringManagerController.java
@@ -1,55 +1,10 @@
-//package com.example.talent_api.controllers;
-//
-//import org.springframework.web.bind.annotation.DeleteMapping;
-//import org.springframework.web.bind.annotation.GetMapping;
-//import org.springframework.web.bind.annotation.PathVariable;
-//import org.springframework.web.bind.annotation.PostMapping;
-//import org.springframework.web.bind.annotation.PutMapping;
-//import org.springframework.web.bind.annotation.RequestMapping;
-//import org.springframework.web.bind.annotation.RestController;
-//
-//@RestController
-//@RequestMapping("/manager")
-//public class HiringManagerController {
-//
-//    //Can change what is returning later
-//    //For now I have it return a String
-//
-//    @GetMapping
-//    public String getAll() {
-//        return "Get list of record of managers";
-//    }
-//
-//    @GetMapping("/{id}")
-//    public String getManagerByID(@PathVariable("id") long id) {
-//        return "This is ID of single record manager: " + id;
-//    }
-//
-//    @PostMapping
-//    public String addManager() {
-//
-//        return "This is to add Manager";
-//    }
-//
-//    @PutMapping("/{id}")
-//    public String putCustomer(@PathVariable("id") long managerID) {
-//
-//        return "Update this Manager by id: " + managerID;
-//
-//    }
-//
-//    @DeleteMapping("/{id}")
-//    public String deleteCustomerById(@PathVariable("id") long id) {
-//        // repo.delete(id);
-//        return "Delete manager by this ID";
-//    }
-//
-//}
 package com.example.talent_api.controllers;
 
 import com.example.talent_api.entity.HiringManager;
 import com.example.talent_api.repository.HiringManagerRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -62,35 +17,46 @@ public class HiringManagerController {
     private HiringManagerRepository hiringManagerRepository;
 
     @GetMapping
-    public List<HiringManager> getAllHiringManagers() {
-        return (List<HiringManager>) hiringManagerRepository.findAll();
+    public ResponseEntity<List<HiringManager>> getAllHiringManagers() {
+        return new ResponseEntity<>((List<HiringManager>) hiringManagerRepository.findAll(), HttpStatus.OK);
     }
 
     @GetMapping("/{id}")
-    public HiringManager getHiringManagerById(@PathVariable Long id) {
-        return hiringManagerRepository.findById(id).orElse(null);
+    public ResponseEntity<HiringManager> getHiringManagerById(@PathVariable Long id) {
+        HiringManager hiringManager = hiringManagerRepository.findById(id).orElse(null);
+        if (hiringManager != null) {
+            return new ResponseEntity<>(hiringManager, HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
     }
 
     @PostMapping
-    public HiringManager createHiringManager(@RequestBody HiringManager hiringManager) {
-        return hiringManagerRepository.save(hiringManager);
+    public ResponseEntity<HiringManager> createHiringManager(@RequestBody HiringManager hiringManager) {
+        return new ResponseEntity<>(hiringManagerRepository.save(hiringManager), HttpStatus.CREATED);
     }
 
     @PutMapping("/{id}")
-    public HiringManager updateHiringManager(@PathVariable Long id, @RequestBody HiringManager hiringManagerDetails) {
+    public ResponseEntity<HiringManager> updateHiringManager(@PathVariable Long id, @RequestBody HiringManager hiringManagerDetails) {
         HiringManager hiringManager = hiringManagerRepository.findById(id).orElse(null);
         if (hiringManager != null) {
             hiringManager.setName(hiringManagerDetails.getName());
             hiringManager.setEmail(hiringManagerDetails.getEmail());
             hiringManager.setDepartment(hiringManagerDetails.getDepartment());
             hiringManager.setPhone(hiringManagerDetails.getPhone());
-            return hiringManagerRepository.save(hiringManager);
+            return new ResponseEntity<>(hiringManagerRepository.save(hiringManager), HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
-        return null;
     }
 
     @DeleteMapping("/{id}")
-    public void deleteHiringManager(@PathVariable Long id) {
-        hiringManagerRepository.deleteById(id);
+    public ResponseEntity<Void> deleteHiringManager(@PathVariable Long id) {
+        if (hiringManagerRepository.existsById(id)) {
+            hiringManagerRepository.deleteById(id);
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        } else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
     }
 }
